requirements.txt
# FOBOH News Aggregation & LangChain RAG System v4.0.0
# Core dependencies for LangChain + GPT-4o + FAISS architecture

# Web Framework & Server
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
jinja2==3.1.2
starlette==0.27.0

# Environment & Configuration
python-dotenv==1.0.0
pydantic==2.5.0

# HTTP & Async
aiohttp==3.9.1
httpx==0.25.2
aiofiles==23.2.1

# LangChain & AI (Primary Stack)
langchain==0.1.0
langchain-openai==0.0.5
langchain-community==0.0.10
openai==1.6.1

# Vector Database & Embeddings
faiss-cpu==1.7.4

# Web Scraping & Data Processing
beautifulsoup4==4.12.2
requests==2.31.0
feedparser==6.0.10
newspaper3k==0.2.8
selenium==4.16.2
webdriver-manager==4.0.1

# Legacy AI Support (ChromaDB backup)
sentence-transformers==2.2.2
chromadb==0.4.18
transformers==4.36.2

# NLP & Text Processing
spacy==3.7.2
nltk==3.8.1
scikit-learn==1.3.2

# Database & Storage
sqlalchemy==2.0.23
alembic==1.13.1

# Data Processing & Utilities
pandas==2.1.4
numpy==1.25.2
python-dateutil==2.8.2
pytz==2023.3
schedule==1.2.0

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Optional: Redis for caching (if needed)
redis==5.0.1 